# [BE] Toward Metal Iterator (step 2)